requirements.txt
streamlit==1.28.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.1.78
streamlit-drawable-canvas==0.9.3
scikit-image==0.21.0
matplotlib==3.7.2